Consolidate the test infrastructure between presto-native and presto-on-spark
1. Refactor presto-native and presto-on-spark unit tests to allow them to share
common functionalities (properties, util functions etc).
2. Simplified the TestPrestoSparkNativeExecution to allow adding new test query
by just adding one line.
3. Add test query results verification compared with the results from vanilla
Presto
